Adds basic bitcoin style proof of work distrubted logging
This adds a set of classes to handle the semantics of operating a proof
of work style distributed log. Currently missing some of the edge
cases for handling out of order messages and such. Does not
and will not handle the networking side of this.

Currently one supports byte size difficulty increments